requirements.txt
numpy
pandas
opencv-python
easyocr
pillow
pdf2image
python-dateutil
streamlit
matplotlib
plotly
pdfplumber
